backend/machine_learning/requirements.txt
pandas
tensorflow
scikit-learn
odfpy
